FilterButton에서만 필터기능 완
result 변수를 이용해 필터링 된 어레이 넘김.
